// next-front/src/types/daiaryData.ts
export type DiaryData = {
  id: number;
  actionName: string;
  actionDetail: string;
  originality: string;
  getSomething: string;
  questionOne: string;
  answerOne: string;
  questionTwo: string;
  answerTwo: string;
  imgUrl: string;
  isBestDiary: number;
  startTime: string;
  endTime: string;
  date: string;
};